refactor(runtime): split interface implementations into separate modules
